# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CircleCue API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/scenarios": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "List Scenarios",
        "operationId": "list_scenarios_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Create Scenario",
        "operationId": "create_scenario_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{scenario_id}": {
      "patch": {
        "tags": [
          "scenarios"
        ],
        "summary": "Update Scenario",
        "operationId": "update_scenario_scenarios__scenario_id__patch",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scenarios"
        ],
        "summary": "Delete Scenario",
        "operationId": "delete_scenario_scenarios__scenario_id__delete",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{scenario_id}/dry-run": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Dry Run Scenario",
        "operationId": "dry_run_scenario_scenarios__scenario_id__dry_run_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dev/tick": {
      "post": {
        "tags": [
          "dev"
        ],
        "summary": "Dev Tick Post",
        "operationId": "dev_tick_post_dev_tick_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TickRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "dev"
        ],
        "summary": "Dev Tick Get",
        "operationId": "dev_tick_get_dev_tick_get",
        "parameters": [
          {
            "name": "minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Minutes"
            }
          },
          {
            "name": "seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Seconds"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityStateTrigger": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "activity_state",
            "title": "Kind"
          },
          "activity_type": {
            "type": "string",
            "title": "Activity Type"
          },
          "elapsed_min": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Elapsed Min",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "kind",
          "activity_type"
        ],
        "title": "ActivityStateTrigger"
      },
...
      "Availability": {
        "properties": {
          "calls": {
            "$ref": "#/components/schemas/Calls",
            "default": "ok"
          },
          "messages": {
            "$ref": "#/components/schemas/Messages",
            "default": "ok"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "Availability"
      },
      "BatteryTrigger": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "battery",
            "title": "Kind"
          },
          "threshold_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Threshold Pct"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "threshold_pct"
        ],
        "title": "BatteryTrigger"
      },
      "Calls": {
        "type": "string",
        "enum": [
          "ok",
          "prefer_not",
          "no"
        ],
        "title": "Calls"
      },
...
      "CreateReminderEffect": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "create_reminder",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "title"
        ],
        "title": "CreateReminderEffect"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualTrigger": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "manual",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "ManualTrigger"
      },
...
      "Messages": {
        "type": "string",
        "enum": [
          "ok",
          "later"
        ],
        "title": "Messages"
      },
...
      "NotifyEffect": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "notify",
            "title": "Kind"
          },
          "notification_kind": {
            "type": "string",
            "title": "Notification Kind"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "notification_kind"
        ],
        "title": "NotifyEffect"
      },
...
      "ScenarioCondition": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "operator": {
            "type": "string",
            "enum": [
              "eq",
              "neq",
              "gte",
              "lte",
              "contains"
            ],
            "title": "Operator"
          },
          "value": {
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "field",
          "operator",
          "value"
        ],
        "title": "ScenarioCondition"
      },
...
      "ScenarioWriteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "trigger": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TimeTrigger"
              },
              {
                "$ref": "#/components/schemas/ActivityStateTrigger"
              },
              {
                "$ref": "#/components/schemas/BatteryTrigger"
              },
              {
                "$ref": "#/components/schemas/ManualTrigger"
              }
            ],
            "title": "Trigger",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "activity_state": "#/components/schemas/ActivityStateTrigger",
                "battery": "#/components/schemas/BatteryTrigger",
                "manual": "#/components/schemas/ManualTrigger",
                "time": "#/components/schemas/TimeTrigger"
              }
            }
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ScenarioCondition"
            },
            "type": "array",
            "title": "Conditions"
          },
          "effects": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SetActivityEffect"
                },
                {
                  "$ref": "#/components/schemas/SetAvailabilityEffect"
                },
                {
                  "$ref": "#/components/schemas/SuppressNotificationsEffect"
                },
                {
                  "$ref": "#/components/schemas/NotifyEffect"
                },
                {
                  "$ref": "#/components/schemas/CreateReminderEffect"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "create_reminder": "#/components/schemas/CreateReminderEffect",
                  "notify": "#/components/schemas/NotifyEffect",
                  "set_activity": "#/components/schemas/SetActivityEffect",
                  "set_availability": "#/components/schemas/SetAvailabilityEffect",
                  "suppress_notifications": "#/components/schemas/SuppressNotificationsEffect"
                }
              }
            },
            "type": "array",
            "title": "Effects"
          },
          "audience": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Audience"
          },
          "audience_mode": {
            "type": "string",
            "enum": [
              "all_granted",
              "only",
              "except"
            ],
            "title": "Audience Mode",
            "default": "all_granted"
          },
          "notification_rule": {
            "type": "string",
            "enum": [
              "always",
              "first_only",
              "never",
              "only_if_called"
            ],
            "title": "Notification Rule",
            "default": "always"
          },
          "priority": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Priority",
            "default": 50
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "name",
          "trigger",
          "effects"
        ],
        "title": "ScenarioWriteRequest"
      },
      "SetActivityEffect": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "set_activity",
            "title": "Kind"
          },
          "activity_type": {
            "type": "string",
            "title": "Activity Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilitySpec"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "activity_type",
          "title"
        ],
        "title": "SetActivityEffect"
      },
      "SetAvailabilityEffect": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "set_availability",
            "title": "Kind"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "availability"
        ],
        "title": "SetAvailabilityEffect"
      },
...
      "SuppressNotificationsEffect": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "suppress_notifications",
            "title": "Kind"
          },
          "kinds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Kinds"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "SuppressNotificationsEffect"
      },
      "TickRequest": {
        "properties": {
          "minutes": {
            "type": "integer",
            "title": "Minutes",
            "default": 1
          },
          "seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seconds",
            "default": 0
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "type": "object",
        "title": "TickRequest"
      },
...
      "TimeTrigger": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "time",
            "title": "Kind"
          },
          "days": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Days"
          },
          "start_local": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}$",
            "title": "Start Local"
          },
          "end_local": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}$",
            "title": "End Local"
          },
          "week_pattern": {
            "$ref": "#/components/schemas/WeekPattern",
            "default": "every"
          },
          "active_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active From"
          },
          "active_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active To"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "start_local",
          "end_local"
        ],
        "title": "TimeTrigger"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisibilitySpec": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "inherit",
              "private_label",
              "only"
            ],
            "title": "Mode",
            "default": "inherit"
          },
          "label_override": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Override"
          },
          "viewer_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Viewer Ids"
          }
        },
        "type": "object",
        "title": "VisibilitySpec"
      },
      "WeekPattern": {
        "type": "string",
        "enum": [
          "every",
          "A",
          "B"
        ],
        "title": "WeekPattern"
      }
    }
  }
}